Use options for fchown and fchownat arguments
Wraps Uid and Gid arguments as Options so that users can provide
None as a value in place of -1

Closes: #220

+++ src/imp/linux_raw/process/syscalls.rs
@@ -107,7 +107,7 @@ pub(crate) fn getgid() -> Gid {
     #[cfg(not(any(target_arch = "x86", target_arch = "sparc", target_arch = "arm")))]
     unsafe {
         let gid = ret_usize_infallible(syscall_readonly!(__NR_getgid)) as __kernel_gid_t;
-        Gid::from_raw(gid)
+        Gid::from_raw(gid as i32)
     }
 }
 
@@ -122,7 +122,7 @@ pub(crate) fn getegid() -> Gid {
     #[cfg(not(any(target_arch = "x86", target_arch = "sparc", target_arch = "arm")))]
     unsafe {
         let gid = ret_usize_infallible(syscall_readonly!(__NR_getegid)) as __kernel_gid_t;
-        Gid::from_raw(gid)
+        Gid::from_raw(gid as i32)
     }
 }
 
@@ -136,7 +136,7 @@ pub(crate) fn getuid() -> Uid {
     #[cfg(not(any(target_arch = "x86", target_arch = "sparc", target_arch = "arm")))]
     unsafe {
         let uid = ret_usize_infallible(syscall_readonly!(__NR_getuid)) as __kernel_uid_t;
-        Uid::from_raw(uid)
+        Uid::from_raw(uid as i32)
     }
 }
 
@@ -151,7 +151,7 @@ pub(crate) fn geteuid() -> Uid {
     #[cfg(not(any(target_arch = "x86", target_arch = "sparc", target_arch = "arm")))]
     unsafe {
         let uid = ret_usize_infallible(syscall_readonly!(__NR_geteuid)) as __kernel_uid_t;
-        Uid::from_raw(uid)
+        Uid::from_raw(uid as i32)
     }
 }
 
@@ -227,7 +227,7 @@ pub(crate) fn getpriority_user(uid: Uid) -> io::Result<i32> {
             - ret_c_int(syscall_readonly!(
                 __NR_getpriority,
                 c_uint(linux_raw_sys::general::PRIO_USER),
-                c_uint(uid.as_raw())
+                c_int(uid.as_raw())
             ))?)
     }
 }
@@ -262,7 +262,7 @@ pub(crate) fn setpriority_user(uid: Uid, priority: i32) -> io::Result<()> {
         ret(syscall_readonly!(
             __NR_setpriority,
             c_uint(linux_raw_sys::general::PRIO_USER),
-            c_uint(uid.as_raw()),
+            c_int(uid.as_raw()),
             c_int(priority)
         ))
     }
